working hit & miss damage instructions with tests, also changed i32 -> i16 for damage

+++ src/damage_calc.rs
@@ -170,45 +170,45 @@ fn volatile_status_modifier(
     return modifier;
 }
 
-fn get_damage_rolls(damage: f32, damage_roll_type: DamageRolls) -> Vec<i32> {
+fn get_damage_rolls(damage: f32, damage_roll_type: DamageRolls) -> Vec<i16> {
     match damage_roll_type {
         DamageRolls::Min => {
-            return vec![(damage * 0.85) as i32];
+            return vec![(damage * 0.85) as i16];
         }
         DamageRolls::Average => {
-            return vec![(damage * 0.925) as i32];
+            return vec![(damage * 0.925) as i16];
         }
         DamageRolls::Max => {
-            return vec![damage as i32];
+            return vec![damage as i16];
         }
         DamageRolls::MinMax => {
-            return vec![(damage * 0.85) as i32, damage as i32];
+            return vec![(damage * 0.85) as i16, damage as i16];
         }
         DamageRolls::MinMaxAverage => {
             return vec![
-                (damage * 0.85) as i32,
-                (damage * 0.925) as i32,
-                damage as i32,
+                (damage * 0.85) as i16,
+                (damage * 0.925) as i16,
+                damage as i16,
             ];
         }
         DamageRolls::All => {
             return vec![
-                (damage * 0.85) as i32,
-                (damage * 0.86) as i32,
-                (damage * 0.87) as i32,
-                (damage * 0.88) as i32,
-                (damage * 0.89) as i32,
-                (damage * 0.90) as i32,
-                (damage * 0.91) as i32,
-                (damage * 0.92) as i32,
-                (damage * 0.93) as i32,
-                (damage * 0.94) as i32,
-                (damage * 0.95) as i32,
-                (damage * 0.96) as i32,
-                (damage * 0.97) as i32,
-                (damage * 0.98) as i32,
-                (damage * 0.99) as i32,
-                damage as i32,
+                (damage * 0.85) as i16,
+                (damage * 0.86) as i16,
+                (damage * 0.87) as i16,
+                (damage * 0.88) as i16,
+                (damage * 0.89) as i16,
+                (damage * 0.90) as i16,
+                (damage * 0.91) as i16,
+                (damage * 0.92) as i16,
+                (damage * 0.93) as i16,
+                (damage * 0.94) as i16,
+                (damage * 0.95) as i16,
+                (damage * 0.96) as i16,
+                (damage * 0.97) as i16,
+                (damage * 0.98) as i16,
+                (damage * 0.99) as i16,
+                damage as i16,
             ];
         }
     }
@@ -224,7 +224,7 @@ pub fn calculate_damage(
     attacking_side: SideReference,
     choice: &Choice,
     damage_rolls: DamageRolls,
-) -> Option<Vec<i32>> {
+) -> Option<Vec<i16>> {
     if choice.base_power <= 0.0 {
         return None;
     }
